Recreated methods to get the correct data from Riot Servers, now the server should correctly respond with the last detailed information of the last match played

+++ backend/src/main/java/com/example/codecoolers_rift/model/Match.java
@@ -35,13 +35,13 @@ public class Match {
     @JsonProperty("lane")
     public String lane;
     @JsonProperty("gameId")
-    public Integer gameId;
+    public Long gameId;
     @JsonProperty("champion")
     public Integer champion;
     @JsonProperty("platformId")
     public String platformId;
     @JsonProperty("timestamp")
-    public Integer timestamp;
+    public Long timestamp;
     @JsonProperty("queue")
     public Integer queue;
     @JsonProperty("role")
